Finished Day 6 & 7 of Advent 2016
Pretty Nifty stuff. Day6, found out that defaultdict wasn't as powerful
as I thought. Needs some investigation but it looks like it might have
actually only created 1 default dict that was copied by reference 6
times in the list.

Day 7 was just a lot of pain. This is definitely not the most efficient
solution with a 1 pass. I thought it was much clearer though. We have
two different behaviors depending on the type of sequence, but they go
through the exact same processing.